commands: move checkout code to checkout.rs
