+++ cli/src/commands/mod.rs
@@ -18,6 +18,7 @@ mod backout;
 mod bench;
 mod branch;
 mod cat;
+mod checkout;
 mod debug;
 mod git;
 mod operation;
@@ -81,7 +82,7 @@ enum Commands {
     Branch(branch::BranchSubcommand),
     #[command(alias = "print")]
     Cat(cat::CatArgs),
-    Checkout(CheckoutArgs),
+    Checkout(checkout::CheckoutArgs),
     Chmod(ChmodArgs),
     Commit(CommitArgs),
     #[command(subcommand)]
@@ -251,23 +252,6 @@ struct ConfigSetArgs {
 struct ConfigEditArgs {
     #[clap(flatten)]
     pub config_args: ConfigArgs,
-}
-
-/// Create a new, empty change and edit it in the working copy
-///
-/// For more information, see
-/// https://github.com/martinvonz/jj/blob/main/docs/working-copy.md.
-#[derive(clap::Args, Clone, Debug)]
-#[command(visible_aliases = &["co"])]
-struct CheckoutArgs {
-    /// The revision to update to
-    revision: RevisionArg,
-    /// Ignored (but lets you pass `-r` for consistency with other commands)
-    #[arg(short = 'r', hide = true)]
-    unused_revision: bool,
-    /// The change description to use
-    #[arg(long = "message", short, value_name = "MESSAGE")]
-    message_paragraphs: Vec<String>,
 }
 
 /// Stop tracking specified paths in the working copy
@@ -1348,30 +1332,6 @@ fn cmd_config_edit(
 ) -> Result<(), CommandError> {
     let config_path = get_new_config_file_path(&args.config_args.get_source_kind(), command)?;
     run_ui_editor(command.settings(), &config_path)
-}
-
-#[instrument(skip_all)]
-fn cmd_checkout(
-    ui: &mut Ui,
-    command: &CommandHelper,
-    args: &CheckoutArgs,
-) -> Result<(), CommandError> {
-    let mut workspace_command = command.workspace_helper(ui)?;
-    let target = workspace_command.resolve_single_rev(&args.revision, ui)?;
-    let mut tx =
-        workspace_command.start_transaction(&format!("check out commit {}", target.id().hex()));
-    let commit_builder = tx
-        .mut_repo()
-        .new_commit(
-            command.settings(),
-            vec![target.id().clone()],
-            target.tree_id().clone(),
-        )
-        .set_description(cli_util::join_message_paragraphs(&args.message_paragraphs));
-    let new_commit = commit_builder.write()?;
-    tx.edit(&new_commit).unwrap();
-    tx.finish(ui)?;
-    Ok(())
 }
 
 #[instrument(skip_all)]
@@ -3886,7 +3846,7 @@ pub fn run_command(ui: &mut Ui, command_helper: &CommandHelper) -> Result<(), Co
         Commands::Version(sub_args) => cmd_version(ui, command_helper, sub_args),
         Commands::Init(sub_args) => cmd_init(ui, command_helper, sub_args),
         Commands::Config(sub_args) => cmd_config(ui, command_helper, sub_args),
-        Commands::Checkout(sub_args) => cmd_checkout(ui, command_helper, sub_args),
+        Commands::Checkout(sub_args) => checkout::cmd_checkout(ui, command_helper, sub_args),
         Commands::Untrack(sub_args) => cmd_untrack(ui, command_helper, sub_args),
         Commands::Files(sub_args) => cmd_files(ui, command_helper, sub_args),
         Commands::Cat(sub_args) => cat::cmd_cat(ui, command_helper, sub_args),
